backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
requests
typing_extensions
google-generative-ai
